Project - Remove conditionals for deploying to Debian or RedHat systems
It's a sensible enough idea since mostly only the installation and
configuration steps are different. However it adds a layer of complexity
and more importantly there's simply not the time to find out whether
dual-system playbooks are a good idea or not. It's probably simpler to
have one playbook per platform since that eliminates the conditionals
and either a free to evolve as needed.